requirements.txt
praw
pandas
matplotlib
seaborn
nltk
textblob
vaderSentiment
python-dotenv
tabulate